+++ src/core/events/update-status-and-appearance.js
@@ -1,8 +1,8 @@
 import {forkJoin} from "rxjs";
 
-import updateAutoSaveStatus$ from "../status/update-auto-save";
-import updatePriceUpdateStatus$ from "../status/update-price-update";
-import updateExtensionAppearance$ from "../update-extension-appearance";
+import updateAutoSaveStatus$ from "./status/update-auto-save";
+import updatePriceUpdateStatus$ from "./status/update-price-update";
+import updateExtensionAppearance$ from "./update-extension-appearance";
 
 function updateStatusAndAppearance(currentURL, domain, forcePageTrackingTo, url) {
     return forkJoin(